added feature to check if user exists by id

+++ src/main/java/com/eletronicos/jfctecnologia/controllers/UserController.java
@@ -40,8 +40,14 @@ public class UserController {
     @DeleteMapping("delete/{id}")
     @Transactional
     public ResponseEntity<Void> deleteUser(@PathVariable Long id) {
-        userService.deleteUser(id);
-        return ResponseEntity.noContent().build();
+
+        if (userService.existsById(id)) {
+            userService.deleteUser(id);
+            return ResponseEntity.noContent().build();
+        } else {
+
+            return ResponseEntity.notFound().build();
+        }
     }
 
 }
